requirements.txt
# Quest API测试依赖包
requests>=2.25.1
PyYAML>=5.4.1
pytest>=6.2.5
pytest-html>=3.1.1
pytest-metadata>=1.11.0